requirements.txt
Django==4.2.0
Pillow==10.0.0
django-crispy-forms==2.0
django-crispy-bootstrap5==2025.6